refactor: Modularize the request logging middleware
